tests/functional/requirements.txt
aiohttp==3.8.3
aioredis==1.3.1
asyncpg==0.27.0
backoff==2.2.1
orjson==3.8.8
psycopg2-binary==2.9.5
pydantic==1.10.4
pytest==7.2.1
pytest_asyncio==0.20.3
python-dotenv==0.21.0